Refactor wait time handling in waiting list join flow
Changed wait time parameter from long to int throughout the client-server waiting list join process for consistency. Added logging and improved response handling on the server to return the full waitlist order object upon successful join.

+++ Alpha_version/Client/src/logic/WaitingListController.java
@@ -125,7 +125,7 @@ public class WaitingListController {
 		client.handleMessageFromClientUI(new Message(Api.ASK_WAITING_LIST_CHECK_AVAILABILITY, dinersAmount));
 	}
 	
-	public void joinWaitingList(int diners, long waitTimeMinutes) {
+	public void joinWaitingList(int diners, int waitTimeMinutes) {
 	    Map<String, Object> details = new HashMap<>();
 	    details.put("diners", diners);
 	    details.put("waitTime", waitTimeMinutes);
